Add vagrantfile
Adds a vagrantfile to compile the Go and Vibe.d versions on a VirtualBox
environment with 2048MB of RAM and 4 CPUs. It also includes `wrk` on the
VM for benchmarking.

To compile D, it installs GDC, which is the GCC frontend for D, and LDC,
which is the LLVM frontend for D.